+++ src/cz/tul/dic/engine/opencl/solvers/AbstractTaskSolver.java
@@ -171,7 +171,7 @@ public abstract class AbstractTaskSolver extends Observable {
         final CorrelationResult result;
         if (r1 == null) {
             result = r2;
-        } else if (Float.compare(r1.getValue(), r2.getValue()) == 0) {
+        } else if (Double.compare(r1.getValue(), r2.getValue()) == 0) {
             result = DeformationUtils.getAbs(r1.getDeformation()) < DeformationUtils.getAbs(r2.getDeformation()) ? r1 : r2;
         } else {
             result = r1.getValue() > r2.getValue() ? r1 : r2;
